Fee cache test added
